Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-canvas-tainting-video-with-range-request.https-expected.txt


FAIL range responses from single origin (same-origin) assert_equals: expected "NOT_TAINTED" but got "LOAD_ERROR"
FAIL range responses from multiple origins (cross-origin first) assert_equals: expected "TAINTED" but got "LOAD_ERROR"
FAIL range responses from single origin with both opaque and non-opaque responses assert_equals: expected "TAINTED" but got "LOAD_ERROR"
